Extract parsing code to separate component
Reduce the size of `trimutil.cpp` by extracting out the parsing code to
`parser.h`/`parser.cpp`.
