+++ src/trimutil.cpp
@@ -25,620 +25,18 @@
 #include "depsreader.h"
 #include "graph.h"
 #include "logreader.h"
-#include "murmur_hash.h"
+#include "parser.h"
 
-#include <ninja/eval_env.h>
-#include <ninja/lexer.h>
 #include <ninja/util.h>
 
-#include <array>
 #include <cassert>
-#include <filesystem>
-#include <forward_list>
 #include <fstream>
 #include <iostream>
 #include <numeric>
-#include <set>
-#include <sstream>
-#include <stdexcept>
-#include <string>
-#include <unordered_map>
-#include <vector>
 
 namespace trimja {
 
 namespace {
-
-struct TransparentHash {
-  using is_transparent = void;
-  std::size_t operator()(const std::string& v) const {
-    return std::hash<std::string_view>{}(v);
-  }
-  std::size_t operator()(std::string_view v) const {
-    return std::hash<std::string_view>{}(v);
-  }
-};
-
-class BasicScope {
-  std::unordered_map<std::string, std::string, TransparentHash, std::equal_to<>>
-      m_variables;
-
- public:
-  BasicScope() = default;
-
-  std::string_view set(std::string_view key, std::string&& value) {
-    // `operator[]` does not support `is_transparent` and `emplace` may or
-    // may not move from `value` so this is the best way to avoid allocations
-    return m_variables.emplace(key, "").first->second = std::move(value);
-  }
-
-  bool appendValue(std::string& output, std::string_view name) const {
-    const auto it = m_variables.find(name);
-    if (it == m_variables.end()) {
-      return false;
-    } else {
-      output += it->second;
-      return true;
-    }
-  }
-};
-
-struct Rule {
-  static const std::size_t bindingCount = 11;
-
-  std::string name;
-  std::array<unsigned char, bindingCount + 1>
-      lookup;  // TODO: pack into uint64_t
-  std::vector<EvalString> bindings;
-
-  explicit Rule(std::string_view name) : name(name), lookup(), bindings() {
-    lookup.fill(std::numeric_limits<unsigned char>::max());
-  }
-
-  std::size_t getLookupIndex(std::string_view varName) const {
-    const std::array<std::string_view, bindingCount> names = {
-        "command",
-        "depfile",
-        "dyndep",
-        "description",
-        "deps",
-        "generator",
-        "pool",
-        "restat",
-        "rspfile",
-        "rspfile_content",
-        "msvc_deps_prefix",
-    };
-    return std::find(names.begin(), names.end(), varName) - names.begin();
-  }
-
-  bool add(std::string_view varName, EvalString&& value) {
-    const std::size_t lookupIndex = getLookupIndex(varName);
-    if (lookupIndex >= bindingCount) {
-      return false;
-    }
-    const std::size_t bindingIndex = lookup[lookupIndex];
-    if (bindingIndex < bindings.size()) {
-      bindings[bindingIndex] = std::move(value);
-    } else {
-      bindings.push_back(std::move(value));
-      lookup[lookupIndex] = static_cast<unsigned char>(bindings.size() - 1);
-    }
-    return true;
-  }
-
-  const EvalString* lookupVar(std::string_view varName) const {
-    const std::size_t bindingIndex = lookup[getLookupIndex(varName)];
-    return bindingIndex < bindings.size() ? &bindings[bindingIndex] : nullptr;
-  }
-};
-
-template <typename SCOPE>
-struct EdgeScope {
-  std::span<const std::string> ins;
-  std::span<const std::string> outs;
-  BasicScope local;
-  SCOPE& parent;
-  const Rule& rule;
-
-  EdgeScope(SCOPE& parent,
-            const Rule& rule,
-            std::span<const std::string> ins,
-            std::span<const std::string> outs)
-      : ins(ins), outs(outs), local(), parent(parent), rule(rule) {}
-
-  bool appendValue(std::string& output, std::string_view name) const {
-    // From https://ninja-build.org/manual.html#ref_scope
-    // Variable declarations indented in a build block are scoped to the build
-    // block. The full lookup order for a variable expanded in a build block (or
-    // the rule is uses) is:
-    //   1. Special built-in variables ($in, $out).
-    //   2. Build-level variables from the build block.
-    //   3. Rule-level variables from the rule block (i.e. $command). (Note from
-    //      the above discussion on expansion that these are expanded "late",
-    //      and may make use of in-scope bindings like $in.)
-    //   4. File-level variables from the file that the build line was in.
-    //   5. Variables from the file that included that file using the subninja
-    //      keyword.
-    if (name == "in") {
-      appendPaths(output, ins, ' ');
-      return true;
-    } else if (name == "out") {
-      appendPaths(output, outs, ' ');
-      return true;
-    } else if (name == "in_newline") {
-      appendPaths(output, ins, '\n');
-      return true;
-    } else {
-      if (local.appendValue(output, name)) {
-        return true;
-      }
-
-      if (const EvalString* value = rule.lookupVar(name)) {
-        evaluate(output, *value, *this);
-        return true;
-      }
-
-      return parent.appendValue(output, name);
-    }
-  }
-
- private:
-  static void appendPaths(std::string& output,
-                          std::span<const std::string> paths,
-                          const char separator) {
-    auto it = paths.begin();
-    const auto end = paths.end();
-    if (it == end) {
-      return;
-    }
-
-    goto skipSeparator;
-    for (; it != end; ++it) {
-      output += separator;
-    skipSeparator:
-      appendEscapedString(output, *it);
-    }
-  }
-};
-
-template <typename SCOPE>
-void evaluate(std::string& output,
-              const EvalString& variable,
-              const SCOPE& scope) {
-  for (const auto& [string, type] : variable.parsed_) {
-    if (type == EvalString::RAW) {
-      output += string;
-    } else {
-      scope.appendValue(output, string);
-    }
-  }
-}
-
-struct BuildCommand {
-  enum Resolution {
-    // Print the entire build command
-    Print,
-
-    // Create a phony command for all the (implicit) outputs
-    Phony,
-  };
-
-  Resolution resolution = Phony;
-
-  // The location of our entire build command inside `m_parts`
-  std::size_t partsIndex = std::numeric_limits<std::size_t>::max();
-
-  // The hash of the build command
-  std::uint64_t hash = 0;
-
-  // Map each output index to the string containing the
-  // "build out1 out$ 2 | implicitOut3" (note no newline and no trailing `|`
-  // or `:`)
-  std::string_view outStr;
-
-  // Map each output index to the string containing the validation edges
-  // e.g. "|@ validation1 validation2" (note no newline and no leading
-  // space)
-  std::string_view validationStr;
-};
-
-struct BuildContext {
-  // An optional storage for the contents for the input ninja files.  This is
-  // useful when processing `include` and `subninja` and we need to extend the
-  // lifetime of the file contents until all parsing has finished
-  std::forward_list<std::string> fileContents;
-
-  // All parts of the input build file, indexing into an element of
-  // `fileContents`.  Note that we may swap out build command sections with
-  // phony commands.
-  std::vector<std::string_view> parts;
-
-  // All build commands and default statements mentioned
-  std::vector<BuildCommand> commands;
-
-  // Map each output index to the index within `command`.  Use -1 for a
-  // value that isn't an output to a build command (i.e. a source file)
-  std::vector<std::size_t> nodeToCommand;
-
-  // Our rules keyed by name
-  std::unordered_map<std::string, Rule, TransparentHash, std::equal_to<>> rules;
-
-  // Our top-level variables
-  BasicScope fileScope;
-
-  // Our graph
-  Graph graph;
-
-  BuildContext() = default;
-};
-
-class Parser {
-  BuildContext* m_ctx;
-
-  // Our lexer
-  Lexer m_lexer;
-
- public:
-  std::size_t getPathIndex(std::string_view path) {
-    const std::size_t index = m_ctx->graph.addPath(path);
-    if (index >= m_ctx->nodeToCommand.size()) {
-      m_ctx->nodeToCommand.resize(index + 1,
-                                  std::numeric_limits<std::size_t>::max());
-    }
-    return index;
-  }
-
-  std::size_t getDefault() {
-    const std::size_t index = m_ctx->graph.addDefault();
-    if (index >= m_ctx->nodeToCommand.size()) {
-      m_ctx->nodeToCommand.resize(index + 1,
-                                  std::numeric_limits<std::size_t>::max());
-    }
-    return index;
-  }
-
- private:
-  void expectToken(Lexer::Token expected) {
-    const Lexer::Token token = m_lexer.ReadToken();
-    if (token != expected) {
-      std::stringstream msg;
-      msg << "Expected " << Lexer::TokenName(expected) << " but got "
-          << Lexer::TokenName(token) << '\0';
-      throw std::runtime_error(msg.view().data());
-    }
-  }
-
-  void skipPool() {
-    if (!m_lexer.SkipIdent()) {
-      throw std::runtime_error("Missing name for pool");
-    }
-
-    expectToken(Lexer::NEWLINE);
-    while (m_lexer.PeekToken(Lexer::INDENT)) {
-      skipLet();
-    }
-  }
-
-  void parseInclude() {
-    EvalString pathEval;
-    std::string err;
-    if (!m_lexer.ReadPath(&pathEval, &err)) {
-      throw std::runtime_error(err);
-    }
-    std::string path;
-    evaluate(path, pathEval, m_ctx->fileScope);
-
-    if (!std::filesystem::exists(path)) {
-      throw std::runtime_error(
-          std::format("Unable to find included file {}!", path));
-    }
-
-    std::stringstream ninjaCopy;
-    std::ifstream ninja(path);
-    ninjaCopy << ninja.rdbuf();
-    m_ctx->fileContents.push_front(ninjaCopy.str());
-
-    Parser subParser;
-    subParser.parse(*m_ctx, path, m_ctx->fileContents.front());
-
-    expectToken(Lexer::NEWLINE);
-  }
-
-  void skipRule() {
-    std::string_view name;
-    if (!m_lexer.ReadIdent(&name)) {
-      throw std::runtime_error("Missing name for rule");
-    }
-
-    const auto [it, inserted] = m_ctx->rules.emplace(name, name);
-    if (!inserted) {
-      std::stringstream ss;
-      ss << "Duplicate rule '" << name << "' found!" << '\0';
-      throw std::runtime_error(ss.view().data());
-    }
-
-    Rule& rule = it->second;
-
-    expectToken(Lexer::NEWLINE);
-    while (m_lexer.PeekToken(Lexer::INDENT)) {
-      std::string_view key;
-      EvalString value;
-      parseLet(key, value);
-      if (!rule.add(key, std::move(value))) {
-        std::stringstream ss;
-        ss << "Unexpected variable '" << key << "' in rule '" << name
-           << "' found!" << '\0';
-        throw std::runtime_error(ss.view().data());
-      }
-    }
-  }
-
-  void skipLet() {
-    if (!m_lexer.SkipIdent()) {
-      throw std::runtime_error("Missing variable name");
-    }
-
-    std::string err;
-    expectToken(Lexer::EQUALS);
-    if (!m_lexer.SkipVarValue(&err)) {
-      throw std::runtime_error(err);
-    }
-  }
-
-  void parseLet(std::string_view& key, EvalString& value) {
-    if (!m_lexer.ReadIdent(&key)) {
-      throw std::runtime_error("Missing variable name");
-    }
-
-    expectToken(Lexer::EQUALS);
-
-    std::string err;
-    if (!m_lexer.ReadVarValue(&value, &err)) {
-      throw std::runtime_error(err);
-    }
-  }
-
-  template <typename SCOPE>
-  std::size_t collectPaths(std::vector<std::string>& result,
-                           SCOPE& scope,
-                           std::string* err) {
-    EvalString out;
-    std::size_t count = 0;
-    while (true) {
-      out.Clear();
-      if (!m_lexer.ReadPath(&out, err)) {
-        throw std::runtime_error(*err);
-      }
-      if (out.empty()) {
-        break;
-      }
-
-      std::string& output = result.emplace_back();
-      evaluate(output, out, scope);
-      [[maybe_unused]] std::uint64_t slashBits;
-      CanonicalizePath(&output, &slashBits);
-      ++count;
-    }
-
-    return count;
-  }
-
-  void handleEdge(const char* start) {
-    std::vector<std::string> outs;
-    std::string errStorage;
-    std::string* err = &errStorage;
-
-    {
-      EvalString out;
-      if (!m_lexer.ReadPath(&out, err)) {
-        throw std::runtime_error(*err);
-      }
-      while (!out.empty()) {
-        // TODO: Allow bindings from the rule to be looked up on edges
-        std::string& output = outs.emplace_back();
-        evaluate(output, out, m_ctx->fileScope);
-        [[maybe_unused]] std::uint64_t slashBits;
-        CanonicalizePath(&output, &slashBits);
-        out.Clear();
-        if (!m_lexer.ReadPath(&out, err)) {
-          throw std::runtime_error(*err);
-        }
-      }
-    }
-    const std::size_t outSize = outs.size();
-
-    if (m_lexer.PeekToken(Lexer::PIPE)) {
-      // Collect implicit outs
-      collectPaths(outs, m_ctx->fileScope, err);
-    }
-
-    if (outs.empty()) {
-      throw std::runtime_error("Missing output paths in build command");
-    }
-
-    expectToken(Lexer::COLON);
-
-    // Mark the outputs for later
-    const std::string_view outStr(start, m_lexer.position());
-
-    std::string_view ruleName;
-    if (!m_lexer.ReadIdent(&ruleName)) {
-      throw std::runtime_error("Missing rule name for build command");
-    }
-
-    const Rule& rule = [&] {
-      const auto ruleIt = m_ctx->rules.find(ruleName);
-      if (ruleIt == m_ctx->rules.end()) {
-        throw std::runtime_error("Unable to find " + std::string(ruleName) +
-                                 " rule");
-      }
-      return ruleIt->second;
-    }();
-
-    // Collect inputs
-    std::vector<std::string> ins;
-    const std::size_t inSize = collectPaths(ins, m_ctx->fileScope, err);
-
-    // Collect implicit inputs
-    if (m_lexer.PeekToken(Lexer::PIPE)) {
-      collectPaths(ins, m_ctx->fileScope, err);
-    }
-
-    // Collect build-order dependencies
-    if (m_lexer.PeekToken(Lexer::PIPE2)) {
-      collectPaths(ins, m_ctx->fileScope, err);
-    }
-
-    // Collect validations but ignore what they are. If we include a build
-    // command it will include the validation.  If that validation has a
-    // required input then we include that, otherwise the validation is
-    // `phony`ed out.
-    std::string_view validationStr;
-    const char* validationStart = m_lexer.position();
-    if (m_lexer.PeekToken(Lexer::PIPEAT)) {
-      std::vector<std::string> validations;
-      collectPaths(validations, m_ctx->fileScope, err);
-      validationStr = std::string_view(validationStart, m_lexer.position());
-    }
-
-    expectToken(Lexer::NEWLINE);
-
-    EdgeScope scope(m_ctx->fileScope, rule, std::span(ins.data(), inSize),
-                    std::span(outs.data(), outSize));
-
-    EvalString value;
-    while (m_lexer.PeekToken(Lexer::INDENT)) {
-      std::string_view key;
-      parseLet(key, value);
-      std::string result;
-      evaluate(result, value, scope);
-      scope.local.set(key, std::move(result));
-      value.Clear();
-    }
-
-    const std::size_t partsIndex = m_ctx->parts.size();
-    m_ctx->parts.emplace_back(start, m_lexer.position());
-
-    // Add the build command
-    const std::size_t commandIndex = m_ctx->commands.size();
-    BuildCommand& buildCommand = m_ctx->commands.emplace_back();
-    buildCommand.partsIndex = partsIndex;
-    buildCommand.validationStr = validationStr;
-    buildCommand.outStr = outStr;
-
-    // Add outputs to the graph and link to the build command
-    std::vector<std::size_t> outIndices;
-    for (const std::string& out : outs) {
-      const std::size_t outIndex = getPathIndex(out);
-      outIndices.push_back(outIndex);
-      m_ctx->nodeToCommand[outIndex] = commandIndex;
-    }
-
-    // Add inputs to the graph and add the edges to the graph
-    for (const std::string& in : ins) {
-      const std::size_t inIndex = getPathIndex(in);
-      for (const std::size_t outIndex : outIndices) {
-        m_ctx->graph.addEdge(inIndex, outIndex);
-      }
-    }
-
-    {
-      std::string command;
-      scope.appendValue(command, "command");
-      std::string rspcontent;
-      scope.appendValue(rspcontent, "rspfile_content");
-      if (!rspcontent.empty()) {
-        command += ";rspfile=";
-        command += rspcontent;
-      }
-      buildCommand.hash = murmur_hash::hash(command.data(), command.size());
-    }
-  }
-
-  void handleDefault(const char* start) {
-    std::vector<std::string> ins;
-    std::string err;
-    collectPaths(ins, m_ctx->fileScope, &err);
-    if (ins.empty()) {
-      throw std::runtime_error("Expected path");
-    }
-
-    expectToken(Lexer::NEWLINE);
-
-    const std::size_t partsIndex = m_ctx->parts.size();
-    m_ctx->parts.emplace_back(start, m_lexer.position());
-
-    const std::size_t commandIndex = m_ctx->commands.size();
-    BuildCommand& buildCommand = m_ctx->commands.emplace_back();
-    buildCommand.partsIndex = partsIndex;
-
-    const std::size_t outIndex = getDefault();
-    m_ctx->nodeToCommand[outIndex] = commandIndex;
-    for (const std::string& in : ins) {
-      m_ctx->graph.addEdge(getPathIndex(in), outIndex);
-    }
-  }
-
- public:
-  Parser() = default;
-
-  void parse(BuildContext& ctx,
-             const std::filesystem::path& filename,
-             std::string_view input) {
-    m_ctx = &ctx;
-    m_lexer.Start(filename.string(), input);
-
-    while (true) {
-      const char* start = m_lexer.position();
-      const Lexer::Token token = m_lexer.ReadToken();
-      switch (token) {
-        case Lexer::POOL:
-          skipPool();
-          m_ctx->parts.emplace_back(start, m_lexer.position());
-          break;
-        case Lexer::BUILD:
-          handleEdge(start);
-          break;
-        case Lexer::RULE:
-          skipRule();
-          m_ctx->parts.emplace_back(start, m_lexer.position());
-          break;
-        case Lexer::DEFAULT:
-          handleDefault(start);
-          break;
-        case Lexer::IDENT: {
-          m_lexer.UnreadToken();
-          std::string_view key;
-          EvalString value;
-          parseLet(key, value);
-          std::string result;
-          evaluate(result, value, m_ctx->fileScope);
-          m_ctx->fileScope.set(key, std::move(result));
-          m_ctx->parts.emplace_back(start, m_lexer.position());
-        } break;
-        case Lexer::INCLUDE:
-          parseInclude();
-          break;
-        case Lexer::SUBNINJA:
-          throw std::runtime_error("subninja not yet supported");
-        case Lexer::ERROR:
-          throw std::runtime_error("Parsing error");
-        case Lexer::TEOF: {
-          return;
-        }
-        case Lexer::NEWLINE:
-          break;
-        default: {
-          std::stringstream msg;
-          msg << "Unexpected token " << Lexer::TokenName(token) << '\0';
-          throw std::runtime_error(msg.view().data());
-        }
-      }
-    }
-    throw std::logic_error("Not reachable");
-  }
-};
 
 enum Requirement : char {
   // This build command is needed only by `default` and has no inputs marked
@@ -696,7 +94,7 @@ void markInputsAsRequired(Graph& graph,
 
 void parseDepFile(const std::filesystem::path& ninjaDeps,
                   Graph& graph,
-                  Parser& parser) {
+                  BuildContext& ctx) {
   std::ifstream deps(ninjaDeps);
   DepsReader reader(deps);
   std::vector<std::size_t> lookup;
@@ -710,7 +108,7 @@ void parseDepFile(const std::filesystem::path& ninjaDeps,
         }
         // No need to normalize the path here as it is normalized when it
         // is written to the `.ninja_deps` file
-        lookup[view.index] = parser.getPathIndex(view.path);
+        lookup[view.index] = ctx.getPathIndex(view.path);
         break;
       }
       case 1: {
@@ -764,16 +162,14 @@ void parseLogFile(const std::filesystem::path& ninjaLog,
 
 void TrimUtil::trim(std::ostream& output,
                     const std::filesystem::path& ninjaFile,
-                    const std::string& ninjaFileContents,
+                    std::string_view ninjaFileContents,
                     std::istream& changed) {
   BuildContext ctx;
 
-  // Add all our of changes files to the graph and mark them as required
-  Graph& graph = ctx.graph;
-
   // Parse the build file
-  Parser parser;
-  parser.parse(ctx, ninjaFile, ninjaFileContents);
+  ParserUtil::parse(ctx, ninjaFile, ninjaFileContents);
+
+  Graph& graph = ctx.graph;
 
   const std::filesystem::path builddir = [&] {
     std::string path;
@@ -786,7 +182,7 @@ void TrimUtil::trim(std::ostream& output,
   // Add all dynamic dependencies from `.ninja_deps` to the graph
   if (const std::filesystem::path ninjaDeps = builddir / ".ninja_deps";
       std::filesystem::exists(ninjaDeps)) {
-    parseDepFile(ninjaDeps, graph, parser);
+    parseDepFile(ninjaDeps, graph, ctx);
   }
 
   std::vector<Requirement> requirements(graph.size(), Requirement::CreatePhony);
